refactored query and tests
