+++ server/database/models/Organization.js
@@ -16,10 +16,6 @@ const organizationSchema = new Schema(
       enum: constants.database.ORGANIZATIONS_TYPE,
       required: true,
     },
-    verified: {
-      type: Boolean,
-      default: true,
-    },
     phoneNumber: String,
     email: {
       type: String,
