--- server/database/queries/organization.js
@@ -1,9 +1,0 @@
-const Organization = require("./../models/Organization");
-
-module.exports.findByName = name => Organization.findOne({ name });
-
-module.exports.addNew = ({ name, category, verified }) => Organization.create({
-  name,
-  category,
-  verified,
-});
