+++ server/database/queries/organizations/addOrganization.js
@@ -1,0 +1,8 @@
+const Organization = require("../../models/Organization");
+
+module.exports.findByName = name => Organization.findOne({ name });
+
+module.exports.addNew = ({ name, category }) => Organization.create({
+  name,
+  category,
+});
