+++ server/database/queries/organizations/index.js
@@ -1,5 +1,6 @@
 const Organization = require("./../../models/Organization");
 const getOrgsByCategory = require("./getOrgsByCategory");
+const { addNew, findByName } = require("./addOrganization");
 
 module.exports.getOrgsByCategory = getOrgsByCategory;
 
@@ -8,3 +9,7 @@ module.exports.deleteOrganization = id => Organization.deleteOne({ _id: id });
 module.exports.getOrganizationById = id => Organization.findById(id);
 
 module.exports.updateOrgsById = (id, data) => Organization.findByIdAndUpdate(id, { $set: data });
+
+module.exports.addNew = addNew;
+
+module.exports.findByName = findByName;
